refactor scheduler addon p1 to p5 -squashed
refactor scheduler ass addon p1
refactor 3
refator p5
