+++ Win10_BrightnessSlider/Gui/_Form1.cs
@@ -207,9 +207,6 @@ v1.8.23
         }
 
         private IKeyboardMouseEvents m_GlobalHook;
-        private System.Windows.Forms.Timer schedulerTimer;
-        private DateTime lastScheduleCheck = DateTime.MinValue;
-
         private void Form1_Load(object sender, EventArgs e)
         {
             eSetVis(false);
@@ -255,17 +252,35 @@ v1.8.23
             SetColors_themeAware_RightClickMenus();
             GUI_Update__CtxMenu_isRunAtStartup();
 
-            //GUI_Update_StatesOnControls(); //also set CtxMenu Run At StartUp
+			//GUI_Update_StatesOnControls(); //also set CtxMenu Run At StartUp
 
-            // Initialize scheduler AFTER monitors are ready
-            InitializeScheduler();
+			// Initialize scheduler AFTER monitors are ready
+			//InitializeScheduler();
+			init_Scheduler_ADDON();
 
-            notifyIcon_bright.Text = "Win10_BrightnessSlider";
+
+			notifyIcon_bright.Text = "Win10_BrightnessSlider";
         }
 
+		private void init_Scheduler_ADDON()
+		{
+			try
+			{
+				var scheduler = new z_Schedule.SchedulerAddon(this);
+				scheduler.Initialize();
+
+				tbxLog_AppendText("\r\n Scheduler ADDON initialized...");
+			}
+			catch (Exception ex)
+			{
+				// Show error to screen so we know why it failed
+				MessageBox.Show("Scheduler Addon Error:\n" + ex.Message + "\n\n" + ex.StackTrace);
+				RamLogger.Log("Scheduler ADDON Error: " + ex.Message);
+			}
+		}
 
 
-        private void wire_events()
+		private void wire_events()
         {
             tbxLog_AppendText("\r\n wire_events - mglobal_hook ... ");
             //clicked outside windows--  deactivate doesnt work every time.
@@ -894,335 +909,8 @@ v1.8.23
         }
 
 
-        private ToolStripMenuItem CreateScheduledBrightnessMenuItem()
-        {
-            var settings = Settings_json.Get();
-            var mi_scheduler = new ToolStripMenuItem("Auto-Dim Schedule");
-            mi_scheduler.Checked = settings.EnableScheduledBrightness;
-            mi_scheduler.ToolTipText = "Automatically adjust brightness at scheduled times";
-
-            mi_scheduler.Click += (snd, ev) =>
-            {
-                var _mi = snd as ToolStripMenuItem;
-                _mi.Checked = !_mi.Checked;
-
-                Settings_json.Update(st =>
-                {
-                    st.EnableScheduledBrightness = _mi.Checked;
-
-                    // Add default schedules if none exist
-                    if (_mi.Checked && (st.BrightnessSchedules == null || st.BrightnessSchedules.Count == 0))
-                    {
-                        st.BrightnessSchedules = new List<BrightnessSchedule>
-                        {
-                            new BrightnessSchedule { Time = "23:00", BrightnessPercent = 25, Enabled = true },  // 11 PM → 25%
-                            new BrightnessSchedule { Time = "07:00", BrightnessPercent = 100, Enabled = true }  // 7 AM → 100%
-                        };
-                    }
-                });
-
-                if (_mi.Checked)
-                {
-                    MessageBox.Show(
-                        "Scheduled brightness enabled!\n\n" +
-                        "Default schedules:\n" +
-                        "• 11:00 PM → 25% (dim for night)\n" +
-                        "• 7:00 AM → 100% (bright for morning)\n\n" +
-                        "Use 'Edit Schedules' to customize times and brightness levels.",
-                        "Auto-Dim Schedule",
-                        MessageBoxButtons.OK,
-                        MessageBoxIcon.Information);
-                }
-            };
-
-            return mi_scheduler;
-        }
-
-        private ToolStripMenuItem CreateEditSchedulesMenuItem()
-        {
-            var mi_editSchedules = new ToolStripMenuItem("Edit Schedules...");
-            mi_editSchedules.ToolTipText = "Open schedule editor to customize brightness schedules";
-
-            mi_editSchedules.Click += (snd, ev) =>
-            {
-                using (var editor = new Gui.ScheduleEditor())
-                {
-                    editor.ShowDialog();
-                }
-            };
-
-            return mi_editSchedules;
-        }
-
-        private void InitializeScheduler()
-        {
-            try
-            {
-                FileLogger.Log("InitializeScheduler: Starting...");
-
-                schedulerTimer = new System.Windows.Forms.Timer();
-                schedulerTimer.Interval = 60000; // Check every minute
-                schedulerTimer.Tick += SchedulerTimer_Tick;
-                schedulerTimer.Start();
-
-                FileLogger.Log("InitializeScheduler: Timer started, now calling CheckAndApplySchedulesAtStartup");
-
-                // Check schedules immediately on startup and apply current brightness if needed
-                CheckAndApplySchedulesAtStartup();
-
-                FileLogger.Log("InitializeScheduler: Completed successfully");
-            }
-            catch (Exception ex)
-            {
-                FileLogger.Log("InitializeScheduler error: " + ex.Message + "\n" + ex.StackTrace);
-            }
-        }
-
-        private void SchedulerTimer_Tick(object sender, EventArgs e)
-        {
-            CheckAndApplySchedules();
-        }
-
-        private void CheckAndApplySchedules()
-        {
-            try
-            {
-                var settings = Settings_json.Get();
-                if (!settings.EnableScheduledBrightness || settings.BrightnessSchedules == null)
-                    return;
-
-                var now = DateTime.Now;
-                var currentTime = now.TimeOfDay;
-                var currentMinute = new TimeSpan(now.Hour, now.Minute, 0);
-
-                // Only check once per minute to avoid duplicate applications
-                if (lastScheduleCheck.Hour == now.Hour && lastScheduleCheck.Minute == now.Minute)
-                    return;
-
-                lastScheduleCheck = now;
-
-                // Check each schedule
-                foreach (var schedule in settings.BrightnessSchedules.Where(s => s.Enabled && s.AppliesToday()))
-                {
-                    var scheduleTime = schedule.GetTimeSpan();
-                    var scheduleMinute = new TimeSpan(scheduleTime.Hours, scheduleTime.Minutes, 0);
-
-                    if (scheduleMinute == currentMinute)
-                    {
-                        // Apply this schedule
-                        FileLogger.Log($"Applying scheduled brightness: {schedule.GetBrightnessString()} at {schedule.Time} ({schedule.GetDaysString()})");
-                        ApplyScheduledBrightnessFromSchedule(schedule);
-                        break; // Only apply one schedule per minute
-                    }
-                }
-            }
-            catch (Exception ex)
-            {
-                FileLogger.Log("CheckAndApplySchedules error: " + ex.Message);
-            }
-        }
-
-        private void CheckAndApplySchedulesAtStartup()
-        {
-            try
-            {
-                FileLogger.Log("CheckAndApplySchedulesAtStartup: Starting...");
-
-                var settings = Settings_json.Get();
-                FileLogger.Log($"CheckAndApplySchedulesAtStartup: EnableScheduledBrightness={settings.EnableScheduledBrightness}, ScheduleCount={settings.BrightnessSchedules?.Count ?? 0}");
-
-                if (!settings.EnableScheduledBrightness)
-                {
-                    FileLogger.Log("CheckAndApplySchedulesAtStartup: Scheduler is disabled");
-                    return;
-                }
-
-                if (settings.BrightnessSchedules == null || settings.BrightnessSchedules.Count == 0)
-                {
-                    FileLogger.Log("CheckAndApplySchedulesAtStartup: No schedules configured");
-                    return;
-                }
-
-                var now = DateTime.Now;
-                var currentTime = now.TimeOfDay;
-                var yesterday = now.AddDays(-1).DayOfWeek;
-                FileLogger.Log($"CheckAndApplySchedulesAtStartup: Current time is {now:HH:mm:ss} ({now.DayOfWeek})");
-
-                // Find the most recent schedule that should be active
-                BrightnessSchedule mostRecentSchedule = null;
-                TimeSpan? mostRecentTime = null;
-                bool foundTodaySchedule = false;
-
-                // First, check for schedules that have already passed TODAY
-                foreach (var schedule in settings.BrightnessSchedules.Where(s => s.Enabled))
-                {
-                    var appliesToday = schedule.AppliesToday();
-                    var scheduleTime = schedule.GetTimeSpan();
-                    FileLogger.Log($"CheckAndApplySchedulesAtStartup: Checking TODAY schedule {schedule.Time} → {schedule.BrightnessPercent}% - AppliesToday={appliesToday}, ScheduleTime={scheduleTime}, CurrentTime={currentTime}");
-
-                    if (!appliesToday)
-                        continue;
-
-                    // Check if this schedule has already passed today
-                    if (scheduleTime <= currentTime)
-                    {
-                        FileLogger.Log($"CheckAndApplySchedulesAtStartup: Schedule {schedule.Time} has passed today");
-                        foundTodaySchedule = true;
-                        // This schedule happened today already
-                        if (mostRecentTime == null || scheduleTime > mostRecentTime.Value)
-                        {
-                            mostRecentTime = scheduleTime;
-                            mostRecentSchedule = schedule;
-                            FileLogger.Log($"CheckAndApplySchedulesAtStartup: This is now the most recent schedule");
-                        }
-                    }
-                }
-
-                // If no schedules passed today yet, check YESTERDAY's schedules
-                if (!foundTodaySchedule)
-                {
-                    FileLogger.Log($"CheckAndApplySchedulesAtStartup: No schedules passed today yet, checking YESTERDAY ({yesterday})");
-
-                    foreach (var schedule in settings.BrightnessSchedules.Where(s => s.Enabled))
-                    {
-                        // Check if this schedule applied yesterday
-                        bool appliedYesterday = schedule.Days == null || schedule.Days.Count == 0 || schedule.Days.Count == 7 || schedule.Days.Contains(yesterday);
-                        var scheduleTime = schedule.GetTimeSpan();
-
-                        FileLogger.Log($"CheckAndApplySchedulesAtStartup: Checking YESTERDAY schedule {schedule.Time} → {schedule.BrightnessPercent}% - AppliedYesterday={appliedYesterday}");
-
-                        if (!appliedYesterday)
-                            continue;
-
-                        // Find the latest schedule from yesterday
-                        if (mostRecentTime == null || scheduleTime > mostRecentTime.Value)
-                        {
-                            mostRecentTime = scheduleTime;
-                            mostRecentSchedule = schedule;
-                            FileLogger.Log($"CheckAndApplySchedulesAtStartup: YESTERDAY schedule {schedule.Time} is now the most recent");
-                        }
-                    }
-                }
-
-                // If we found a schedule that should be in effect, apply it
-                if (mostRecentSchedule != null)
-                {
-                    FileLogger.Log($"Startup: Applying scheduled brightness from {mostRecentSchedule.Time}: {mostRecentSchedule.GetBrightnessString()} ({mostRecentSchedule.GetDaysString()})");
-
-                    // Delay slightly to ensure monitors are fully ready
-                    System.Threading.Thread.Sleep(500);
-
-                    ApplyScheduledBrightnessFromSchedule(mostRecentSchedule);
-                }
-                else
-                {
-                    FileLogger.Log("CheckAndApplySchedulesAtStartup: No applicable schedule found for current time");
-                }
-            }
-            catch (Exception ex)
-            {
-                FileLogger.Log("CheckAndApplySchedulesAtStartup error: " + ex.Message + "\n" + ex.StackTrace);
-            }
-        }
-
-        private void ApplyScheduledBrightness(int brightness)
-        {
-            // Legacy method - applies same brightness to all monitors
-            ApplyScheduledBrightnessToAll(brightness);
-        }
-
-        private void ApplyScheduledBrightnessFromSchedule(BrightnessSchedule schedule)
-        {
-            try
-            {
-                if (riScreens == null || riScreens.Count == 0)
-                    return;
-
-                if (schedule.ApplyToAllMonitors)
-                {
-                    // Same brightness for all monitors
-                    ApplyScheduledBrightnessToAll(schedule.BrightnessPercent);
-                }
-                else
-                {
-                    // Per-monitor brightness
-                    ApplyPerMonitorBrightness(schedule);
-                }
-            }
-            catch (Exception ex)
-            {
-                FileLogger.Log("ApplyScheduledBrightnessFromSchedule error: " + ex.Message);
-            }
-        }
-
-        private void ApplyScheduledBrightnessToAll(int brightness)
-        {
-            try
-            {
-                if (riScreens == null || riScreens.Count == 0)
-                    return;
-
-                // Apply to all monitors in parallel
-                var tasks = new List<Task>();
-                foreach (var riScreen in riScreens)
-                {
-                    var screen = riScreen;
-                    tasks.Add(Task.Run(() => screen.SetBrightness(brightness, false)));
-                }
-
-                Task.WaitAll(tasks.ToArray());
-
-                // Update UI on UI thread
-                this.Invoke((Action)(() =>
-                {
-                    GUI_Update__AllSliderControls();
-                    GUI_Update_NotifyIconText();
-                }));
-            }
-            catch (Exception ex)
-            {
-                FileLogger.Log("ApplyScheduledBrightnessToAll error: " + ex.Message);
-            }
-        }
-
-        private void ApplyPerMonitorBrightness(BrightnessSchedule schedule)
-        {
-            try
-            {
-                if (riScreens == null || riScreens.Count == 0 || schedule.PerMonitorBrightness == null)
-                    return;
-
-                var tasks = new List<Task>();
-
-                foreach (var riScreen in riScreens)
-                {
-                    // Find matching brightness entry for this monitor
-                    var monitorId = riScreen.WMIMonitorID?.InstanceName ?? riScreen.dc_TargetDeviceName?.monitorDevicePath;
-                    var entry = schedule.PerMonitorBrightness.FirstOrDefault(e => e.MonitorId == monitorId);
-
-                    int brightness = entry?.BrightnessPercent ?? schedule.BrightnessPercent;
-                    var screen = riScreen;
-
-                    FileLogger.Log($"ApplyPerMonitorBrightness: Setting {screen.avail_MonitorName_clean ?? monitorId} to {brightness}%");
-                    tasks.Add(Task.Run(() => screen.SetBrightness(brightness, false)));
-                }
-
-                Task.WaitAll(tasks.ToArray());
-
-                // Update UI on UI thread
-                this.Invoke((Action)(() =>
-                {
-                    GUI_Update__AllSliderControls();
-                    GUI_Update_NotifyIconText();
-                }));
-            }
-            catch (Exception ex)
-            {
-                FileLogger.Log("ApplyPerMonitorBrightness error: " + ex.Message);
-            }
-        }
-
-        //slider interface, 2 and 3
+       
+        /// --slider interface, 2 and 3
         private List<Iuc_brSlider> getUCSliderList()
         {
             return fLayPnl1.Controls.OfType<Iuc_brSlider>().ToList();
@@ -1821,10 +1509,10 @@ https://github.com/blackholeearth/Win10_BrightnessSlider
                 mi_extras.DropDown.Items.Add(mi_ShowButtons);
                 mi_extras.DropDown.Items.Add(mi_ShowButtons_v2);
                 mi_extras.DropDown.Items.Add(mi_ShowPresetButtons);
-                mi_extras.DropDown.Items.Add("-");
-                mi_extras.DropDown.Items.Add(new ToolStripMenuItem("___Scheduled Brightness___") { Enabled = false });
-                mi_extras.DropDown.Items.Add(CreateScheduledBrightnessMenuItem());
-                mi_extras.DropDown.Items.Add(CreateEditSchedulesMenuItem());
+                //mi_extras.DropDown.Items.Add("-");
+                //mi_extras.DropDown.Items.Add(new ToolStripMenuItem("___Scheduled Brightness___") { Enabled = false });
+                //mi_extras.DropDown.Items.Add(CreateScheduledBrightnessMenuItem());
+                //mi_extras.DropDown.Items.Add(CreateEditSchedulesMenuItem());
                 mi_extras.DropDown.Items.Add("-");
                 mi_extras.DropDown.Items.Add(new ToolStripMenuItem("___ReMap Keys___") { Enabled = false });
                 mi_extras.DropDown.Items.Add(mi_remapKey1);
